requirements.txt
Pillow
gradio
vtracer